✅ fixes annoying test failing locally ➖ removes jsdom dependency
