+++ package.json
@@ -54,7 +54,6 @@
     "jasmine-jquery": "^2.1.1",
     "jest": "^24.9.0",
     "jest-transform-stub": "^2.0.0",
-    "jsdom": "^15.1.0",
     "karma": "^4.2.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^2.0.0",
